google/agents/project_scraper/project_scraper/requirements.txt
google-adk
streamlit
requests
apify-client
python-dotenv

#Para el analizador
google-generativeai